refactor: modularize store.rs into feature-based submodules
