+++ src/tests/store/mod.rs
@@ -1,0 +1,9 @@
+pub mod atomic_tests;
+pub mod builder_tests;
+pub mod concurrency_tests;
+pub mod edge_cases_tests;
+pub mod json_patch_tests;
+pub mod memory_tests;
+pub mod operations_tests;
+pub mod range_tests;
+pub mod timestamp_tests;
